modifié :         src/cvwifiguest.cc 	modifié :         src/cvwifiguest.h
we added one thread that calls, in proactive manner and every 2 seconds, get_winterface_infos(0) function in order to update _list_winterfaces informations.The goal is to get the modified txpower by wconfig for example.
However, the best way is to update _list_winterface only when we detect changing in wireless network interface (todo).