cwifiserver.cc: Bound signal level to appropriate value
If the distance between two clients is low it can result in the signal
level being greater than zero. This doesn't make sense as dBm is
always a negative value. On the client side passing a >0 value to
the kernel as the signal results in extreme signal levels reported
(likely due to a signed/unsigned conversion somewhere).

Check the calculated signal and if greater than zero bound to -10.