Fix incorrect usage of TX_INFO signalling acked frames
Before this change every CMD_FRAME coming from the kernel would be
ack'ed locally through the TX_INFO command. This tells the client
that the frame was acked in all cases, regardless if the server
dropped it.

The kernel does expect a TX_INFO for each CMD_FRAME but the ACK
flag should only be set if the peer received the frame. If the
server drops it the ACK flag should be unset.

To fix this TX_INFO messages can be sent over the "air" just as
frames are. Then clients can choose to include the ACK flag or
not. This did require the server still sends all frames regardless
if they were dropped, but a new boolean was added to signal to
clients if the frame should be dropped. If so the client won't
forward the frame but will still send a TX_INFO (with the ACK
flag set appropriately).